将图层1和2的位置对调一下能占满，现在的思路就是用RGB888单图层，将更改写在了README3.md，但是发现还是花屏。没办法只能看看时钟。这里有了重大发现，XSPI1的时钟竟然开着2分频？操，之前换算下来XSPI1的时钟频率才为42M左右，我去掉了2分频将时钟调为92.31 MHz效果非常不错，但是这里不知道双图层通信会不会也能用，考虑到以后可能UI占图层这里还是只保留1个图层吧，可以以后验证一下

+++ ElectronicWarehouse/Appli/X-CUBE-AI/App/app_x-cube-ai.h
@@ -66,8 +66,8 @@ extern uint8_t *buffer_out_fc;
 
 void MX_X_CUBE_AI_Process_User(ai_result_t *result);
 
-/* Face crop + resize: 800x480 RGB565 display buffer -> 112x112 RGB888 */
-void ai_crop_resize_face_112(uint16_t *src_fb, ai_detection_t *det, uint8_t *output);
+/* Face crop + resize: 800x480 RGB888 display buffer -> 112x112 RGB888 */
+void ai_crop_resize_face_112(uint8_t *src_fb, ai_detection_t *det, uint8_t *output);
 
 /* --- Face Re-ID definitions ------------------------------------------------ */
 #define FACE_EMBEDDING_DIM     128
